Move helper function to helpers.ts
