+++ src/test/setup.ts
@@ -82,16 +82,3 @@ vi.mock('howler', () => ({
 	Howler: mockHowler,
 	Howl: mockHowl,
 }));
-
-/**
- * Create a reusable mock Konva.Image for use in a React ref
- * @returns
- */
-export function createMockKonvaImage() {
-	return {
-		x: () => 0,
-		y: () => 0,
-		// Add any other methods that might be called on the ref
-		setPosition: vi.fn(),
-	};
-}
